+++ webapp/src/components/AuthHeader.js
@@ -1,8 +1,8 @@
 export default function authHeader() {
-  const user = JSON.parse(localStorage.getItem("user"));
-  if (user && user.accessToken) {
-    return { Authorization: "Bearer " + user.accessToken };
-  } else {
-    return {};
-  }
+    const user = JSON.parse(localStorage.getItem("user"));
+    if (user && user.accessToken) {
+        return { Authorization: "Bearer " + user.accessToken };
+    } else {
+        return {};
+    }
 }
