--- webapp/src/components/DeleteUniversity.jsx
@@ -1,12 +1,0 @@
-import React, { Component } from "react";
-
-const DeleteUniversity = () => {
-   
-   
-    return (
-        <div id="DeleteUniversity">
-        <h1> Deleting University</h1>
-    </div>
-    );
-   };
-export default DeleteUniversity;
