+++ webapp/src/components/Explore.jsx
@@ -1,6 +1,7 @@
-import React, { useState, useEffect } from "react";
-import { Form, Button, FormCheck, FormControl, Tab } from "react-bootstrap";
+import React, { useState } from "react";
+import { Form, Button } from "react-bootstrap";
 import { Table } from "react-bootstrap";
+
 const Explore = () => {
     const [Country, setCountry] = useState("");
     const [Course, setCourse] = useState("");
