--- webapp/src/components/StudSignUp.jsx
@@ -1,24 +1,0 @@
-import React, { useState } from "react";
-import {
-    Button,
-    Col,
-    Container,
-    FloatingLabel,
-    Form,
-    Modal,
-    Row,
-} from "react-bootstrap";
-// import { useState } from 'react';
-// import Button from '@material-ui/core/Button';
-// import ModalDialog from './ModalDialog';
-
-const StudSignUp = () => { 
-    const [show, setShow] = useState(false);
-    const handleClose = () => setShow(false);
-    const handleShow = () => setShow(true);
-    return(
-        <h1></h1>
-    );
-    
-};
-export default StudSignUp;
