+++ webapp/src/components/UK.jsx
@@ -15,15 +15,6 @@ const UK = () => {
                     backgroundPosition: "0px",
                 }}
             >
-                {/* <div className="col-3 my-5 mx-2 justify-content-center">
-                    <img
-                        src="https://static.yocket.in/images/universities/countries/uk-logo.jpg"
-                        alt="United Kingdom flag"
-                        sizes="96px"
-                        srcset="https://static.yocket.in/images/universities/countries/uk-logo.jpg 96w"
-                        className="object-cover mt-6 w-24 h-12"
-                    />
-                </div> */}
                 <div className="my-5 mx-2 justify-content-center">
                     <h1 style={{ color: "#6C63FF" }}>
                         <strong>United Kingdom</strong>
